Keep PostgreSQL smoke imports side-effect free
