+++ app/utils/release-calendar.server.ts
@@ -1,6 +1,6 @@
 import { type Prisma } from '@prisma/client'
 import { prisma } from './db.server.ts'
-import { publicTrackingStateWhere } from './lists/visibility.server.ts'
+import { publicTrackingStateWhere } from './lists/visibility.ts'
 import { splitLegacyThumbnail } from './media-detail.ts'
 import { parseStoredNextRelease } from './release-occurrences.server.ts'
